Move dict proxy into its own file
