+++ observ/observables.py
@@ -4,10 +4,12 @@ proxied versions of those datastructures to make them reactive.
 """
 from functools import partial, wraps
 from operator import xor
-import sys
 
 from .dep import Dep
 from .proxy_db import proxy_db
+
+
+TYPE_LOOKUP = {}
 
 
 # TODO: separate file
@@ -296,92 +298,6 @@ def map_traps(obj_cls, traps, trap_map):
         for trap_type, methods in traps.items()
         for method in methods
     }
-
-
-# TODO: separate files for dict, list, set classes
-
-dict_traps = {
-    "READERS": {
-        "copy",
-        "__eq__",
-        "__format__",
-        "__ge__",
-        "__gt__",
-        "__le__",
-        "__len__",
-        "__lt__",
-        "__ne__",
-        "__repr__",
-        "__sizeof__",
-        "__str__",
-        "keys",
-    },
-    "KEYREADERS": {
-        "get",
-        "__contains__",
-        "__getitem__",
-    },
-    "ITERATORS": {
-        "items",
-        "values",
-        "__iter__",
-    },
-    "WRITERS": {
-        "update",
-    },
-    "KEYWRITERS": {
-        "setdefault",
-        "__setitem__",
-    },
-    "DELETERS": {
-        "clear",
-        "popitem",
-    },
-    "KEYDELETERS": {
-        "pop",
-        "__delitem__",
-    },
-}
-
-if sys.version_info >= (3, 8, 0):
-    dict_traps["ITERATORS"].add("__reversed__")
-if sys.version_info >= (3, 9, 0):
-    dict_traps["READERS"].add("__or__")
-    dict_traps["READERS"].add("__ror__")
-    dict_traps["WRITERS"].add("__ior__")
-
-
-# class type(name, bases, dict, **kwds)
-
-
-class DictProxyBase(Proxy):
-    def __init__(self, target, readonly=False, shallow=False):
-        super().__init__(target, readonly=readonly, shallow=shallow)
-
-    def _orphaned_keydeps(self):
-        return set(proxy_db.attrs(self)["keydep"].keys()) - set(self.target.keys())
-
-
-def readonly_dict_proxy__init__(self, target, shallow=False, **kwargs):
-    super(ReadonlyDictProxy, self).__init__(
-        target, shallow=shallow, **{**kwargs, "readonly": True}
-    )
-
-
-DictProxy = type("DictProxy", (DictProxyBase,), map_traps(dict, dict_traps, trap_map))
-ReadonlyDictProxy = type(
-    "ReadonlyDictProxy",
-    (DictProxyBase,),
-    {
-        "__init__": readonly_dict_proxy__init__,
-        **map_traps(dict, dict_traps, trap_map_readonly),
-    },
-)
-
-
-TYPE_LOOKUP = {
-    dict: (DictProxy, ReadonlyDictProxy),
-}
 
 
 # TODO: pair with proxy function
