+++ observ/watcher.py
@@ -13,8 +13,9 @@ from typing import Any, Callable, Optional, TypeVar
 from weakref import ref, WeakSet
 
 from .dep import Dep
+from .dict_proxy import DictProxyBase
 from .list_proxy import ListProxyBase
-from .observables import DictProxyBase, Proxy
+from .observables import Proxy
 from .scheduler import scheduler
 from .set_proxy import SetProxyBase
